+++ carpetmodSrc/carpet/CarpetServer.java
@@ -1,6 +1,7 @@
 package carpet;
 
 import carpet.carpetclient.CarpetClientServer;
+import carpet.helpers.HopperCounter;
 import carpet.helpers.StackTraceDeobfuscator;
 import carpet.helpers.TickSpeed;
 import carpet.helpers.lifetime.LifeTimeTracker;
@@ -125,6 +126,9 @@ public class CarpetServer // static for now - easier to handle all around the co
         TickSpeed.tick(server);
         if (CarpetSettings.redstoneMultimeterLegacy) {
             TickStartEventDispatcher.dispatchEvent(server.getTickCounter());
+        }
+        if (CarpetSettings.cactusCounter || CarpetSettings.hopperCounters != CarpetSettings.HopperCounters.off) {
+            HopperCounter.updateAll(server);
         }
         HUDController.update_hud(server);
         WorldEditBridge.onStartTick();